authentications-management/src/requirements.txt
flask
flask-restful
flask-cors
sqlalchemy
marshmallow-sqlalchemy
marshmallow
python-dotenv
psycopg2-binary
requests
flask-sqlalchemy
pytest
uptrace
flask_jwt_extended